33RULES now v2
hello!  in the time since i uploaded 33RULES.ZIP i found a bunch of pics that were not included, including some previously unreleased ansi — & so this is v2.  it also comes with an intro file, a complete list of packs/groups & an easter egg.  it's better i guess?  could you be so nice as to replace the old one with this (now called 33RULES2.ZIP)?  thanks!  ~33